Move listener classes around
