+++ src/Resources/config/profiler.xml
@@ -7,7 +7,7 @@
     <services>
         <defaults public="false" />
 
-        <service id="Wisembly\AmqpBundle\EventListener\MessagingListener">
+        <service id="Wisembly\AmqpBundle\DataCollector\CollectorInjector">
             <argument type="service">
                 <service class="Wisembly\AmqpBundle\DataCollector\AMQPDataCollector">
                     <tag name="data_collector" template="WisemblyAmqpBundle:Collector:MessageCollector" id="amqp_collector" />
